+++ lib/flow.dart
@@ -1,7 +1,7 @@
 library flow;
 export 'src/flowimpl.dart';
-export 'src/extensions.dart';
+export 'src/extensions.dart' hide DurationX;
 export 'src/builders.dart';
-export 'src/cache.dart';
+export 'src/operators/cache.dart';
 export 'src/exceptions/flow_exception.dart' hide ErrorX;
 export 'src/retries.dart';