+++ lib/src/collectors/flow_collector.dart
@@ -7,6 +7,12 @@
 /// a `Flow`. Concrete implementations must adhere to
 /// this interface and provide a mechanism to send data downstream in the flow.
 abstract class FlowCollector<T> {
+
+  void addError(Object e, [StackTrace? trace]);
+
+
+  void close();
+
   /// Emits a value of type `T` into the flow.
   ///
   /// This method is the core functionality of the `FlowCollector`. Specific