+++ test/cache_test.dart
@@ -2,6 +2,7 @@
 import 'dart:async';
 import 'dart:convert';
 
+import 'package:flow/src/extensions.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flow/flow.dart';
 
@@ -12,7 +13,7 @@ main() {
 
 
   group('FetchOrElseCache', () {
-    test('Test that values from primary flow are emitted when it passes', () {
+    test('Test that values from primary flow are emitted when it passes', ()  async {
       //Arrange
       final cacheFlow = TestCacheFlow<int>(
           'test_key', fromJson: (value) => int.parse('$value'), toJson: (v) => v);
@@ -23,6 +24,7 @@ main() {
 
         const fetchOrElse = FetchOrElseCache();
         await fetchOrElse.handle(cacheFlow, primaryFlow, collector);
+        print('Done First');
       });
 
       //Act and Assert
@@ -53,7 +55,7 @@ main() {
       ]));
     });
 
-    test('Test that when primary doesnt produce values and cache has expired nothing is emitted', () {
+    test('Test that when primary doesnt produce values and cache has expired nothing is emitted', () async {
       final cacheFlow = TestCacheFlow<int>(
           'test_key', fromJson: (value) => int.parse('$value'), toJson: (v) => v);
 
@@ -67,10 +69,12 @@ main() {
         await fetchOrElse.handle(cacheFlow, primaryFlow, collector);
       });
 
-      //Act and Assert
+      // // Act and Assert
       expect(fl.asStream(), emitsInOrder([
         emitsDone
       ]));
+      // fl.collect(print);
+      // await Future.delayed(Duration(seconds: 10));
     });
 
     test('Test that when primary fails and cache has expired the primary error is thrown', () {
@@ -106,7 +110,7 @@ main() {
 
       final fl = flow((collector) async {
         cacheFlow.write(4);
-
+        //
         const cacheOrElseFetch = CacheOrElseFetch();
         await cacheOrElseFetch.handle(cacheFlow, originalFlow, collector);
       });
@@ -274,6 +278,25 @@ main() {
       ]));
     });
   });
+
+  test('Test Cache implementation', () async {
+    final cacheFlow = TestCacheFlow<int>('test_key',
+        fromJson: (value) => int.parse('$value'),
+        toJson: (v) => v
+    );
+
+    cacheFlow.write(2);
+
+    await Future.delayed(10.milliseconds);
+
+    final fl = flow((collector) {
+      collector.emit(1);
+    }).map((value) => value).cache(cacheFlow, CacheThenFetch(maxAge: 100000.milliseconds));
+
+    expect(fl.asStream(), emitsInOrder([
+      2, emitsDone
+    ]));
+  });
 }
 
 class TestCacheFlow<T> extends CacheFlow<T> {
@@ -288,7 +311,8 @@ class TestCacheFlow<T> extends CacheFlow<T> {
   String get _modifiedTimeKey => '${dataKey}_last_modified_time';
 
   @override
-  FutureOr<T?> read() {
+  FutureOr<T?> read() async {
+    await Future.delayed(100.milliseconds);
     final T value = fromJson(sharedPreference[dataKey]);
     return value;
   }